Fixed issue with traversal adding unwanted blocks being added from cluster expansion.
Next up: Work on checking for adjacent island. There must be a better way to check for adjacent islands, it currently adds unwanted water tiles to the board. Also, searching for potential land tiles needs to be improved. Needs to take account how many blocks are left that can be placed and to avoid blocks adjacent to land tiles.