Fixed issue with filling known land.
Improved searching for possible land configurations to include landstems with no parents. Algorithm now can successfully solve land that put in a position where all potential blocks it has access to plus its own blocks are equivlent to the land source's number.

Pretty much means that if it is in a configuration where everything it can possibly touch is guaranteed to be its land, those blocks become its land.